feat(seahaven-cli): introduce `system check` subcommand
- Added a new `check` subcommand to manage system status checks.
- Refactored the existing `system` command to utilize subcommands for better organization.
- Removed the previous dependency checking logic from the `system` command and integrated it into the new `check` command.
- Enhanced error handling for unrecognized subcommands.

This update improves command structure and user experience by clearly separating system management and status checking functionalities.
